extract components from logs page
