+++ src/app/instances/[name]/hardware/components/InstanceTypePage.tsx
@@ -1,10 +1,10 @@
 "use client";
 
-import { useInstanceContext } from "../InstanceContext";
+import { useInstanceContext } from "../../InstanceContext";
 import { useEffect, useState } from "react";
 import axios from "axios";
 import ErrorBanner from "@/app/instances/components/ErrorBanner";
-import SubmissionSpinner from "../../components/SubmissionSpinner";
+import SubmissionSpinner from "../../../components/SubmissionSpinner";
 
 type InstanceTypes = Record<string, string[]>;
