+++ src/app/instances/[name]/hardware/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import { StoragePage } from "./StoragePage";
-import { InstanceTypePage } from "./InstanceTypePage";
+import { StoragePage } from "./components/StoragePage";
+import { InstanceTypePage } from "./components/InstanceTypePage";
 
 export default function HardwarePage() {
   return (
