+++ src/app/instances/[name]/hardware/storageSize.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen, waitFor } from "@testing-library/react";
 import { Suspense } from "react";
-import { StoragePage } from "./StoragePage";
+import { StoragePage } from "./components/StoragePage";
 import { NotificationsContext } from "@/app/NotificationContext";
 
 import axios from "axios";
