+++ src/app/instances/[name]/logs/components/LogsDescription.tsx
@@ -1,0 +1,14 @@
+export default function LogsDescription() {
+  return (
+    <>
+      <h1 className="font-heading1 text-headertext1 text-2xl mb-10">
+        RabbitMQ Logs
+      </h1>
+      <p className="font-text1 text-sm text-pagetext1 mb-6">
+        The following logs are fetched from the RabbitMQ server on your running
+        instance. They are displayed here for convenient debugging. You can also
+        copy your logs as needed.
+      </p>
+    </>
+  )
+}
